Add files via upload
Logistic regression is a statistical method that is commonly used to predict binary outcomes, such as pass /no pass in this case. the model takes the predictors variables (age and physical score) and fits a logistic function to estimate the probability of the binary outcome (pass/no pass).

the accuracy of the model is 0.916,that means correctly predicts the pass or no pass outcome 91.6% of the time .the area under the curve (AUC) plot curve is 0.96 ,which is an indication that the model is performing well.

interpret the result we can look at the coefficients of the logistics regression model . the coefficient of the age indicates how much the probability of passing the audio test changes for a one year increase in age while holding a physical score constant.

overall the logistics regression model can help us understand how age and physical score are related to the likelihood of the hearing high-frequency sounds. and can be used to make predictions about individual participants based on their age and physical score.